Fixed compatibility issue with snircd
Altered how PASS commands are generated, so that if no password is
specified a 1 character value will still be sent in order to not
generate an error from the server that not enough parameters were
specified.

Although the snircd source code as of Wed Jun 23 20:04:45 2010
(http://hg.quakenet.org/snircd/file/37c9c7460603/ircd/m_pass.c) expects
a trailing parameter for the password, it will still be sent in the
standard parameter list as per the official RFCs.