Fixed crashes from a race condition in the socket callbacks and an invalid regex pattern
Fixed crashes from a race condition in the socket which would try to
send data after the connection had been terminated.

Fixed a problem with the regex to process custom colours in messages
that would cause it to fail if reverting to the default colours (i.e. a
single \x03) rather than specifying a new colour.