Extended the favourite servers functionality and fixed an invocation error.
Finished developing the favourite servers functionality so that servers
can be set up as favourites, adding them to the quick connect button and
allowing for commands to be setup that will automatically execute once a
successful connection is made to the network.

Changed IRCMarshal to require the tab host to be specified due to a bug
in which the tab host cannot be found when attempting to invoke it for
processes started on a separate thread.