Added emoticons command, encapsulated the casting of the current tab in MainForm into a property and fixed scrolling issue.
Tabs will now auto-scroll to the bottom when messages are appended while
the tab is not the selected tab.

The selected IRCTabPage can now be accessed via the CurrentTab property
of MainForm. Each time it is accessed it will only re-cast if the
selected tab has been changed since the last cast.